Migrate grafana token code to database package and use grafana verify function in token verification process
