+++ director/director_db.go
@@ -18,18 +18,11 @@
 package director
 
 import (
-	"crypto/rand"
-	"crypto/sha256"
 	"embed"
-	"encoding/hex"
-	"fmt"
-	"strings"
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/jellydator/ttlcache/v3"
 	"github.com/pkg/errors"
-	"golang.org/x/crypto/bcrypt"
 	"gorm.io/gorm"
 	"gorm.io/gorm/logger"
 
@@ -38,16 +31,6 @@ import (
 	"github.com/pelicanplatform/pelican/server_utils"
 )
 
-type GrafanaApiKey struct {
-	ID          string `gorm:"primaryKey;column:id;type:text;not null"`
-	Name        string `gorm:"column:name;type:text"`
-	HashedValue string `gorm:"column:hashed_value;type:text;not null"`
-	Scopes      string `gorm:"column:scopes;type:text"`
-	ExpiresAt   time.Time
-	CreatedAt   time.Time
-	CreatedBy   string `gorm:"column:created_by;type:text"`
-}
-
 type ServerDowntime struct {
 	UUID       string     `gorm:"primaryKey"`
 	Name       string     `gorm:"not null;unique"`
@@ -55,49 +38,6 @@ type ServerDowntime struct {
 	// We don't use gorm default gorm.Model to change ID type to string
 	CreatedAt time.Time
 	UpdatedAt time.Time
-}
-
-var verifiedKeysCache *ttlcache.Cache[string, GrafanaApiKey] = ttlcache.New[string, GrafanaApiKey]()
-
-func VerifyGrafanaApiKey(apiKey string) (bool, error) {
-	parts := strings.Split(apiKey, ".")
-	if len(parts) != 2 {
-		return false, errors.New("invalid API key format")
-	}
-	id := parts[0]
-	secret := parts[1]
-
-	item := verifiedKeysCache.Get(id)
-	if item != nil {
-		cachedToken := item.Value()
-		beforeExpiration := time.Now().Before(cachedToken.ExpiresAt)
-		matches := bcrypt.CompareHashAndPassword([]byte(cachedToken.HashedValue), []byte(secret)) == nil
-		if beforeExpiration && matches {
-			return true, nil
-		}
-	}
-
-	var token GrafanaApiKey
-	//	result := db.First(&token, "id = ?", id)
-	result := database.DirectorDB.First(&token, "id = ?", id)
-	if result.Error != nil {
-		if errors.Is(result.Error, gorm.ErrRecordNotFound) {
-			return false, nil // token not found
-		}
-		return false, errors.Wrap(result.Error, "failed to retrieve the Grafana API key")
-	}
-
-	if time.Now().After(token.ExpiresAt) {
-		return false, nil
-	}
-
-	err := bcrypt.CompareHashAndPassword([]byte(token.HashedValue), []byte(secret))
-	if err != nil {
-		return false, nil
-	}
-
-	verifiedKeysCache.Set(id, token, ttlcache.DefaultTTL)
-	return true, nil
 }
 
 //go:embed migrations/*.sql
@@ -125,57 +65,6 @@ func InitializeDB() error {
 // Shut down the Director's sqlite database
 func shutdownDirectorDB() error {
 	return server_utils.ShutdownDB(database.DirectorDB)
-}
-
-func generateSecret(length int) (string, error) {
-	bytes := make([]byte, length)
-	_, err := rand.Read(bytes)
-	if err != nil {
-		return "", err
-	}
-	return hex.EncodeToString(bytes), nil
-}
-
-func generateTokenID(secret string) string {
-	hash := sha256.Sum256([]byte(secret))
-	return hex.EncodeToString(hash[:])[:5]
-}
-
-func createGrafanaApiKey(name, createdBy, scopes string) (string, error) {
-	expiresAt := time.Now().Add(time.Hour * 24 * 30) // 30 days
-	for {
-		secret, err := generateSecret(32)
-		if err != nil {
-			return "", errors.Wrap(err, "failed to generate a secret")
-		}
-
-		id := generateTokenID(secret)
-
-		hashedValue, err := bcrypt.GenerateFromPassword([]byte(secret), bcrypt.DefaultCost)
-		if err != nil {
-			return "", errors.Wrap(err, "failed to hash the secret")
-		}
-
-		apiKey := GrafanaApiKey{
-			ID:          id,
-			Name:        name,
-			HashedValue: string(hashedValue),
-			Scopes:      scopes,
-			ExpiresAt:   expiresAt,
-			CreatedAt:   time.Now(),
-			CreatedBy:   createdBy,
-		}
-		result := database.DirectorDB.Create(apiKey)
-		if result.Error != nil {
-			isConstraintError := result.Error.Error() == "UNIQUE constraint failed: tokens.id"
-			if !isConstraintError {
-				return "", errors.Wrap(result.Error, "failed to create a new Grafana API key")
-			}
-			// If the ID is already taken, try again
-			continue
-		}
-		return fmt.Sprintf("%s.%s", id, secret), nil
-	}
 }
 
 // Create a new db entry representing the downtime info of a server
